// src/Model/ProfileInformation.ts
import GithubImage from "../Images/githubImage.jpg";

export class ProfileInformation {
  imageUrl: string = GithubImage;
  bio: string = "";
  userName: string = "";
  completeName: string = "";
  repository: number = 0;
  followers: number = 0;
  following: number = 0;
  organisation: string = "";
  website: string = "";
  twitterAccount: string = "";
  dateJoined: string = "";
  location: string = "";
  url: string = "";
}